refactor(ln_voice_over): archive legacy extraction.py under legacy/
The lnvo extract CLI path underperforms the skill-based
/attribute-speakers on today's cheap/local models, but the code
is worth keeping for when local quality catches up. Move the
orchestrator into legacy/ so the active module tree is visually
smaller and the active/legacy split is obvious at a glance.
llm.py and config.MODEL_REGISTRY stay at module root — they're
reusable primitives any future attribution module can import
directly.
