+++ automations/ln_voice_over/legacy/__init__.py
@@ -1,0 +1,4 @@
+"""Legacy modules kept for future use — not part of the active pipeline.
+
+See ``legacy/README.md`` for scope and rationale.
+"""
